Change tilt series type to float in manual background sub.

+++ tomviz/python/Subtract_TiltSer_Background.py
@@ -9,7 +9,9 @@ def transform_scalars(dataset):
     #---------------------------------#
 
     data_bs = utils.get_array(dataset) #get data as numpy array
-    
+
+    data_bs = data_bs.astype(np.float32) #change tilt series type to float
+
     if data_bs is None: #Check if data exists
         raise RuntimeError("No data array found!")
